+++ control-plane/catalog/to-consul/syncer_ent_test.go
@@ -5,8 +5,8 @@ package catalog
 import (
 	"testing"
 
+	"github.com/hashicorp/consul-k8s/control-plane/helper/test"
 	"github.com/hashicorp/consul/api"
-	"github.com/hashicorp/consul/sdk/testutil"
 	"github.com/hashicorp/consul/sdk/testutil/retry"
 	"github.com/stretchr/testify/require"
 )
@@ -15,20 +15,11 @@ import (
 func TestConsulSyncer_ConsulNamespaces(t *testing.T) {
 	t.Parallel()
 
-	a, err := testutil.NewTestServerConfigT(t, nil)
-	require.NoError(t, err)
-	defer a.Stop()
+	testClient := test.TestServerWithMockConnMgrWatcher(t, nil)
+	client := testClient.APIClient
 
-	client, err := api.NewClient(&api.Config{
-		Address: a.HTTPAddr,
-	})
-	require.NoError(t, err)
-
-	s, closer := testConsulSyncerWithConfig(client, func(s *ConsulSyncer) {
+	s, closer := testConsulSyncerWithConfig(testClient, func(s *ConsulSyncer) {
 		s.EnableNamespaces = true
-		s.ConsulNodeServicesClient = &NamespacesNodeServicesClient{
-			Client: client,
-		}
 	})
 	defer closer()
 
@@ -66,20 +57,11 @@ func TestConsulSyncer_ConsulNamespaces(t *testing.T) {
 func TestConsulSyncer_ReapConsulNamespace(t *testing.T) {
 	t.Parallel()
 
-	a, err := testutil.NewTestServerConfigT(t, nil)
-	require.NoError(t, err)
-	defer a.Stop()
+	testClient := test.TestServerWithMockConnMgrWatcher(t, nil)
+	client := testClient.APIClient
 
-	client, err := api.NewClient(&api.Config{
-		Address: a.HTTPAddr,
-	})
-	require.NoError(t, err)
-
-	s, closer := testConsulSyncerWithConfig(client, func(s *ConsulSyncer) {
+	s, closer := testConsulSyncerWithConfig(testClient, func(s *ConsulSyncer) {
 		s.EnableNamespaces = true
-		s.ConsulNodeServicesClient = &NamespacesNodeServicesClient{
-			Client: client,
-		}
 	})
 	defer closer()
 
@@ -135,18 +117,11 @@ func TestConsulSyncer_ReapConsulNamespace(t *testing.T) {
 func TestConsulSyncer_reapServiceInstanceNamespacesEnabled(t *testing.T) {
 	t.Parallel()
 
-	a, err := testutil.NewTestServerConfigT(t, nil)
-	require.NoError(t, err)
-	defer a.Stop()
+	testClient := test.TestServerWithMockConnMgrWatcher(t, nil)
+	client := testClient.APIClient
 
-	client, err := api.NewClient(&api.Config{
-		Address: a.HTTPAddr,
-	})
-	s, closer := testConsulSyncerWithConfig(client, func(s *ConsulSyncer) {
+	s, closer := testConsulSyncerWithConfig(testClient, func(s *ConsulSyncer) {
 		s.EnableNamespaces = true
-		s.ConsulNodeServicesClient = &NamespacesNodeServicesClient{
-			Client: client,
-		}
 	})
 	defer closer()
 
@@ -157,7 +132,7 @@ func TestConsulSyncer_reapServiceInstanceNamespacesEnabled(t *testing.T) {
 	})
 
 	// Create an invalid instance service directly in Consul.
-	_, _, err = client.Namespaces().Create(&api.Namespace{
+	_, _, err := client.Namespaces().Create(&api.Namespace{
 		Name: "foo",
 	}, nil)
 	require.NoError(t, err)